+++ demisto_sdk/__main__.py
@@ -10,72 +10,18 @@ from typing import IO
 
 # Third party packages
 import click
-import dotenv
 import git
 from pkg_resources import get_distribution
 
-from demisto_sdk.commands.common import tools
 from demisto_sdk.commands.common.configuration import Configuration
 # Common tools
 from demisto_sdk.commands.common.constants import FileType
-from demisto_sdk.commands.common.logger import logging_setup
 from demisto_sdk.commands.common.tools import (find_type,
                                                get_last_remote_release_version,
                                                get_release_note_entries,
-                                               print_error, print_warning)
-from demisto_sdk.commands.common.update_id_set import merge_id_sets_from_files
-from demisto_sdk.commands.convert.convert_manager import ConvertManager
-from demisto_sdk.commands.coverage_analyze.coverage_report import \
-    CoverageReport
-from demisto_sdk.commands.create_artifacts.content_artifacts_creator import \
-    ArtifactsManager
-from demisto_sdk.commands.create_id_set.create_id_set import IDSetCreator
-from demisto_sdk.commands.doc_reviewer.doc_reviewer import DocReviewer
-from demisto_sdk.commands.download.downloader import Downloader
-from demisto_sdk.commands.error_code_info.error_code_info import \
-    generate_error_code_information
-from demisto_sdk.commands.find_dependencies.find_dependencies import \
-    PackDependencies
-from demisto_sdk.commands.format.format_module import format_manager
-from demisto_sdk.commands.generate_docs.generate_integration_doc import \
-    generate_integration_doc
-from demisto_sdk.commands.generate_docs.generate_playbook_doc import \
-    generate_playbook_doc
-from demisto_sdk.commands.generate_docs.generate_script_doc import \
-    generate_script_doc
-from demisto_sdk.commands.generate_integration.code_generator import \
-    IntegrationGeneratorConfig
-from demisto_sdk.commands.generate_outputs.generate_outputs import \
-    run_generate_outputs
-from demisto_sdk.commands.generate_test_playbook.test_playbook_generator import \
-    PlaybookTestsGenerator
-from demisto_sdk.commands.init.initiator import Initiator
-from demisto_sdk.commands.integration_diff.integration_diff_detector import \
-    IntegrationDiffDetector
-from demisto_sdk.commands.lint.lint_manager import LintManager
-from demisto_sdk.commands.openapi_codegen.openapi_codegen import \
-    OpenAPIIntegration
-from demisto_sdk.commands.postman_codegen.postman_codegen import \
-    postman_to_autogen_configuration
-# Import demisto-sdk commands
-from demisto_sdk.commands.run_cmd.runner import Runner
-from demisto_sdk.commands.run_playbook.playbook_runner import PlaybookRunner
-from demisto_sdk.commands.secrets.secrets import SecretsValidator
-from demisto_sdk.commands.split.jsonsplitter import JsonSplitter
-from demisto_sdk.commands.split.ymlsplitter import YmlSplitter
-from demisto_sdk.commands.test_content.execute_test_content import \
-    execute_test_content
-from demisto_sdk.commands.unify.generic_module_unifier import \
-    GenericModuleUnifier
-from demisto_sdk.commands.unify.yml_unifier import YmlUnifier
-from demisto_sdk.commands.update_release_notes.update_rn_manager import \
-    UpdateReleaseNotesManager
-from demisto_sdk.commands.update_xsoar_config_file.update_xsoar_config_file import \
-    XSOARConfigFileUpdater
-from demisto_sdk.commands.upload.uploader import ConfigFileParser, Uploader
-from demisto_sdk.commands.validate.validate_manager import ValidateManager
-from demisto_sdk.commands.zip_packs.packs_zipper import (EX_FAIL, EX_SUCCESS,
-                                                         PacksZipper)
+                                               is_external_repository,
+                                               print_error, print_success,
+                                               print_warning)
 
 
 class PathsParamType(click.Path):
@@ -176,6 +122,7 @@ def check_configuration_file(command, args):
 )
 @pass_config
 def main(config, version, release_notes):
+    import dotenv
     dotenv.load_dotenv()  # Load a .env file from the cwd.
     config.configuration = Configuration()
     if not os.getenv('DEMISTO_SDK_SKIP_VERSION_CHECK') or version:  # If the key exists/called to version
@@ -245,6 +192,9 @@ def split(config, **kwargs):
     """Split the code, image and description files from a Demisto integration or script yaml file
     to multiple files(To a package format - https://demisto.pan.dev/docs/package-dir).
     """
+    from demisto_sdk.commands.split.jsonsplitter import JsonSplitter
+    from demisto_sdk.commands.split.ymlsplitter import YmlSplitter
+
     check_configuration_file('split', kwargs)
     file_type: FileType = find_type(kwargs.get('input', ''), ignore_sub_categories=True)
     if file_type not in [FileType.INTEGRATION, FileType.SCRIPT, FileType.GENERIC_MODULE]:
@@ -294,6 +244,7 @@ def split(config, **kwargs):
 @pass_config
 def extract_code(config, **kwargs):
     """Extract code from a Demisto integration or script yaml file."""
+    from demisto_sdk.commands.split.ymlsplitter import YmlSplitter
     check_configuration_file('extract-code', kwargs)
     file_type: FileType = find_type(kwargs.get('input', ''), ignore_sub_categories=True)
     if file_type not in [FileType.INTEGRATION, FileType.SCRIPT]:
@@ -330,17 +281,21 @@ def unify(**kwargs):
     2. GenericModule Unifier - Unifies a GenericModule with its Dashboards to a single JSON object.
      * To use this function - set as input a path to a GenericModule *file*.
     """
-
     check_configuration_file('unify', kwargs)
     # Input is of type Path.
     kwargs['input'] = str(kwargs['input'])
     file_type = find_type(kwargs['input'])
     if file_type == FileType.GENERIC_MODULE:
+        from demisto_sdk.commands.unify.generic_module_unifier import \
+            GenericModuleUnifier
+
         # pass arguments to GenericModule unifier and call the command
         generic_module_unifier = GenericModuleUnifier(**kwargs)
         generic_module_unifier.merge_generic_module_with_its_dashboards()
 
     else:
+        from demisto_sdk.commands.unify.yml_unifier import YmlUnifier
+
         # pass arguments to YML unifier and call the command
         yml_unifier = YmlUnifier(**kwargs)
         yml_unifier.merge_script_package_to_yml()
@@ -364,6 +319,11 @@ def unify(**kwargs):
 @click.option('--zip-all', is_flag=True, help='Zip all the packs in one zip file.', default=False)
 def zip_packs(**kwargs) -> int:
     """Generating zipped packs that are ready to be uploaded to Cortex XSOAR machine."""
+    from demisto_sdk.commands.common.logger import logging_setup
+    from demisto_sdk.commands.upload.uploader import Uploader
+    from demisto_sdk.commands.zip_packs.packs_zipper import (EX_FAIL,
+                                                             EX_SUCCESS,
+                                                             PacksZipper)
     logging_setup(3)
     check_configuration_file('zip-packs', kwargs)
 
@@ -476,6 +436,7 @@ def zip_packs(**kwargs) -> int:
 @pass_config
 def validate(config, **kwargs):
     """Validate your content files. If no additional flags are given, will validated only committed files."""
+    from demisto_sdk.commands.validate.validate_manager import ValidateManager
     check_configuration_file('validate', kwargs)
     sys.path.append(config.configuration.env_dir)
 
@@ -485,7 +446,7 @@ def validate(config, **kwargs):
         print_error('Could not supply the staged flag with the post-commit flag')
         sys.exit(1)
     try:
-        is_external_repo = tools.is_external_repository()
+        is_external_repo = is_external_repository()
         # default validate to -g --post-commit
         if not kwargs.get('validate_all') and not kwargs['use_git'] and not file_path:
             kwargs['use_git'] = True
@@ -561,6 +522,9 @@ def create_content_artifacts(**kwargs) -> int:
        4. content_all - Contains all from content_new and content_test.
        5. uploadable_packs - Contains zipped packs that are ready to be uploaded to Cortex XSOAR machine.
     """
+    from demisto_sdk.commands.common.logger import logging_setup
+    from demisto_sdk.commands.create_artifacts.content_artifacts_creator import \
+        ArtifactsManager
     logging_setup(3)
     check_configuration_file('create-content-artifacts', kwargs)
     artifacts_conf = ArtifactsManager(**kwargs)
@@ -604,6 +568,7 @@ def secrets(config, **kwargs):
     """Run Secrets validator to catch sensitive data before exposing your code to public repository.
      Attach path to whitelist to allow manual whitelists.
      """
+    from demisto_sdk.commands.secrets.secrets import SecretsValidator
     check_configuration_file('secrets', kwargs)
     sys.path.append(config.configuration.env_dir)
     secrets_validator = SecretsValidator(
@@ -665,6 +630,8 @@ def lint(**kwargs):
         Will lookup up what docker image to use and will setup the dev dependencies and file in the target folder.
         If no additional flags specifying the packs are given,will lint only changed files.
     """
+    from demisto_sdk.commands.common.logger import logging_setup
+    from demisto_sdk.commands.lint.lint_manager import LintManager
     logging_setup(verbose=kwargs.get('verbose'),  # type: ignore[arg-type]
                   quiet=kwargs.get('quiet'),  # type: ignore[arg-type]
                   log_path=kwargs.get('log_path'))  # type: ignore[arg-type]
@@ -731,6 +698,8 @@ def lint(**kwargs):
     default='https://storage.googleapis.com/marketplace-dist-dev/code-coverage-reports/coverage-min.json', type=str
 )
 def coverage_analyze(**kwargs):
+    from demisto_sdk.commands.coverage_analyze.coverage_report import \
+        CoverageReport
     try:
         no_degradation_check = kwargs['allowed_coverage_degradation_percentage'] == 100.0
         no_min_coverage_enforcement = kwargs['no_min_coverage_enforcement']
@@ -815,6 +784,7 @@ def format(
     incidenttype/indicatortype/layout/dashboard/classifier/mapper/widget/report file/genericfield/generictype/
     genericmodule/genericdefinition.
     """
+    from demisto_sdk.commands.format.format_module import format_manager
     return format_manager(
         str(input) if input else None,
         str(output) if output else None,
@@ -872,6 +842,9 @@ def upload(**kwargs):
     DEMISTO_API_KEY environment variable should contain a valid Demisto API Key.
     * Note: Uploading classifiers to Cortex XSOAR is available from version 6.0.0 and up. *
     """
+    from demisto_sdk.commands.upload.uploader import ConfigFileParser, Uploader
+    from demisto_sdk.commands.zip_packs.packs_zipper import (EX_FAIL,
+                                                             PacksZipper)
     if kwargs['zip'] or kwargs['input_config_file']:
         if kwargs.pop('zip', False):
             pack_path = kwargs['input']
@@ -934,6 +907,7 @@ def download(**kwargs):
     DEMISTO_BASE_URL environment variable should contain the Demisto server base URL.
     DEMISTO_API_KEY environment variable should contain a valid Demisto API Key.
     """
+    from demisto_sdk.commands.download.downloader import Downloader
     check_configuration_file('download', kwargs)
     downloader: Downloader = Downloader(**kwargs)
     return downloader.download()
@@ -968,6 +942,8 @@ def xsoar_config_file_update(**kwargs):
     DEMISTO_BASE_URL environment variable should contain the Demisto server base URL.
     DEMISTO_API_KEY environment variable should contain a valid Demisto API Key.
     """
+    from demisto_sdk.commands.update_xsoar_config_file.update_xsoar_config_file import \
+        XSOARConfigFileUpdater
     file_updater: XSOARConfigFileUpdater = XSOARConfigFileUpdater(**kwargs)
     return file_updater.update()
 
@@ -1004,6 +980,7 @@ def run(**kwargs):
     DEMISTO_BASE_URL environment variable should contain the Demisto base URL.
     DEMISTO_API_KEY environment variable should contain a valid Demisto API Key.
     """
+    from demisto_sdk.commands.run_cmd.runner import Runner
     check_configuration_file('run', kwargs)
     runner = Runner(**kwargs)
     return runner.run()
@@ -1041,9 +1018,47 @@ def run_playbook(**kwargs):
     Example: DEMISTO_API_KEY=<API KEY> demisto-sdk run-playbook -p 'p_name' -u
     'https://demisto.local'.
     """
+    from demisto_sdk.commands.run_playbook.playbook_runner import \
+        PlaybookRunner
     check_configuration_file('run-playbook', kwargs)
     playbook_runner = PlaybookRunner(**kwargs)
     return playbook_runner.run_playbook()
+
+
+# ====================== run-test-playbook ====================== #
+@main.command()
+@click.help_option(
+    '-h', '--help'
+)
+@click.option(
+    '-tpb', '--test-playbook-path',
+    help="Path to test playbook to run, "
+         "can be a path to specific test playbook or path to pack name for example: Packs/GitHub.",
+    required=False
+)
+@click.option(
+    '--all', is_flag=True,
+    help="Run all the test playbooks from this repository."
+)
+@click.option(
+    '--wait', '-w', is_flag=True, default=True,
+    help="Wait until the test-playbook run is finished and get a response."
+)
+@click.option(
+    '--timeout', '-t',
+    default=90,
+    show_default=True,
+    help="Timeout for the command. The test-playbook will continue to run in your instance"
+)
+@click.option(
+    "--insecure", help="Skip certificate validation.", is_flag=True)
+def run_test_playbook(**kwargs):
+    """Run a test playbooks in your instance."""
+    from demisto_sdk.commands.run_test_playbook.test_playbook_runner import \
+        TestPlaybookRunner
+    check_configuration_file('run-test-playbook', kwargs)
+    test_playbook_runner = TestPlaybookRunner(**kwargs)
+    return test_playbook_runner.manage_and_run_test_playbooks()
 
 
 # ====================== generate-outputs ====================== #
@@ -1103,6 +1118,8 @@ def generate_outputs(**kwargs):
     file/UI/PyCharm. This script auto generates the YAML for a command from the JSON result of the relevant API call
     In addition you can supply examples files and generate the context description directly in the YML from those examples.
     """
+    from demisto_sdk.commands.generate_outputs.generate_outputs import \
+        run_generate_outputs
     check_configuration_file('generate-outputs', kwargs)
     return run_generate_outputs(**kwargs)
 
@@ -1158,6 +1175,8 @@ def generate_outputs(**kwargs):
     "-u", "--upload", help="Whether to upload the test playbook after the generation.", is_flag=True)
 def generate_test_playbook(**kwargs):
     """Generate test playbook from integration or script"""
+    from demisto_sdk.commands.generate_test_playbook.test_playbook_generator import \
+        PlaybookTestsGenerator
     check_configuration_file('generate-test-playbook', kwargs)
     file_type: FileType = find_type(kwargs.get('input', ''), ignore_sub_categories=True)
     if file_type not in [FileType.INTEGRATION, FileType.SCRIPT]:
@@ -1213,6 +1232,7 @@ def init(**kwargs):
     If the script/integration flags are not present, we will create a pack with the given name.
     Otherwise when using the flags we will generate a script/integration based on your selection.
     """
+    from demisto_sdk.commands.init.initiator import Initiator
     check_configuration_file('init', kwargs)
     initiator = Initiator(**kwargs)
     initiator.init()
@@ -1263,6 +1283,12 @@ def init(**kwargs):
     "--skip-breaking-changes", is_flag=True, help="Skip generating of breaking changes section.")
 def generate_docs(**kwargs):
     """Generate documentation for integration, playbook or script from yaml file."""
+    from demisto_sdk.commands.generate_docs.generate_integration_doc import \
+        generate_integration_doc
+    from demisto_sdk.commands.generate_docs.generate_playbook_doc import \
+        generate_playbook_doc
+    from demisto_sdk.commands.generate_docs.generate_script_doc import \
+        generate_script_doc
     check_configuration_file('generate-docs', kwargs)
     input_path: str = kwargs.get('input', '')
     output_path = kwargs.get('output')
@@ -1351,8 +1377,13 @@ def generate_docs(**kwargs):
     help="Fails the process if any duplicates are found.",
     is_flag=True
 )
+@click.option('-mp', '--marketplace', help='The marketplace the id set are created for, that determines which packs are'
+                                           ' inserted to the id set, and which items are present in the id set for '
+                                           'each pack. Default is the XSOAR marketplace, that has all of the packs ',
+              default='xsoar')
 def create_id_set(**kwargs):
     """Create the content dependency tree by ids."""
+    from demisto_sdk.commands.create_id_set.create_id_set import IDSetCreator
     check_configuration_file('create-id-set', kwargs)
     id_set_creator = IDSetCreator(**kwargs)
     id_set_creator.create_id_set()
@@ -1386,6 +1417,8 @@ def create_id_set(**kwargs):
 )
 def merge_id_sets(**kwargs):
     """Merge two id_sets"""
+    from demisto_sdk.commands.common.update_id_set import \
+        merge_id_sets_from_files
     check_configuration_file('merge-id-sets', kwargs)
     first = kwargs['id_set1']
     second = kwargs['id_set2']
@@ -1444,6 +1477,8 @@ def merge_id_sets(**kwargs):
     is_flag=True)
 def update_release_notes(**kwargs):
     """Auto-increment pack version and generate release notes template."""
+    from demisto_sdk.commands.update_release_notes.update_rn_manager import \
+        UpdateReleaseNotesManager
     check_configuration_file('update-release-notes', kwargs)
     if kwargs.get('force') and not kwargs.get('input'):
         print_error('Please add a specific pack in order to force a release notes update.')
@@ -1485,6 +1520,8 @@ def update_release_notes(**kwargs):
               is_flag=True)
 def find_dependencies(**kwargs):
     """Find pack dependencies and update pack metadata."""
+    from demisto_sdk.commands.find_dependencies.find_dependencies import \
+        PackDependencies
     check_configuration_file('find-dependencies', kwargs)
     update_pack_metadata = not kwargs.get('no_update')
     input_path: Path = Path(kwargs["input"])  # To not shadow python builtin `input`
@@ -1559,6 +1596,11 @@ def postman_codegen(
         package: bool
 ):
     """Generates a Cortex XSOAR integration given a Postman collection 2.1 JSON file."""
+    from demisto_sdk.commands.common.logger import logging_setup
+    from demisto_sdk.commands.postman_codegen.postman_codegen import \
+        postman_to_autogen_configuration
+    from demisto_sdk.commands.split.ymlsplitter import YmlSplitter
+
     if verbose:
         logger = logging_setup(verbose=3)
     else:
@@ -1611,6 +1653,9 @@ def generate_integration(input: IO, output: Path, verbose: bool):
     """Generates a Cortex XSOAR integration from a config json file,
     which is generated by commands like postman-codegen
     """
+    from demisto_sdk.commands.common.logger import logging_setup
+    from demisto_sdk.commands.generate_integration.code_generator import \
+        IntegrationGeneratorConfig
     if verbose:
         logging_setup(verbose=3)
 
@@ -1656,6 +1701,8 @@ def openapi_codegen(**kwargs):
     In the first run of the command, an integration configuration file is created, which can be modified.
     Then, the command is run a second time with the integration configuration to generate the actual integration files.
     """
+    from demisto_sdk.commands.openapi_codegen.openapi_codegen import \
+        OpenAPIIntegration
     check_configuration_file('openapi-codegen', kwargs)
     if not kwargs.get('output_dir'):
         output_dir = os.getcwd()
@@ -1667,10 +1714,10 @@ def openapi_codegen(**kwargs):
         try:
             os.mkdir(output_dir)
         except Exception as err:
-            tools.print_error(f'Error creating directory {output_dir} - {err}')
+            print_error(f'Error creating directory {output_dir} - {err}')
             sys.exit(1)
     if not os.path.isdir(output_dir):
-        tools.print_error(f'The directory provided "{output_dir}" is not a directory')
+        print_error(f'The directory provided "{output_dir}" is not a directory')
         sys.exit(1)
 
     input_file = kwargs['input_file']
@@ -1713,7 +1760,7 @@ def openapi_codegen(**kwargs):
     integration.load_file()
     if not kwargs.get('config_file'):
         integration.save_config(integration.configuration, output_dir)
-        tools.print_success(f'Created configuration file in {output_dir}')
+        print_success(f'Created configuration file in {output_dir}')
         if not kwargs.get('use_default', False):
             config_path = os.path.join(output_dir, f'{base_name}_config.json')
             command_to_run = f'demisto-sdk openapi-codegen -i "{input_file}" -cf "{config_path}" -n "{base_name}" ' \
@@ -1731,9 +1778,9 @@ def openapi_codegen(**kwargs):
             sys.exit(0)
 
     if integration.save_package(output_dir):
-        tools.print_success(f'Successfully finished generating integration code and saved it in {output_dir}')
+        print_success(f'Successfully finished generating integration code and saved it in {output_dir}')
     else:
-        tools.print_error(f'There was an error creating the package in {output_dir}')
+        print_error(f'There was an error creating the package in {output_dir}')
         sys.exit(1)
 
 
@@ -1784,6 +1831,8 @@ def test_content(**kwargs):
     run test playbook on the created investigation using mock if possible.
     Collect the result and give a report.
     """
+    from demisto_sdk.commands.test_content.execute_test_content import \
+        execute_test_content
     check_configuration_file('test-content', kwargs)
     execute_test_content(**kwargs)
 
@@ -1823,6 +1872,7 @@ def test_content(**kwargs):
 )
 def doc_review(**kwargs):
     """Check the spelling in .md and .yml files as well as review release notes"""
+    from demisto_sdk.commands.doc_reviewer.doc_reviewer import DocReviewer
     doc_reviewer = DocReviewer(
         file_path=kwargs.get('input'),
         known_words_file_path=kwargs.get('known_words'),
@@ -1859,7 +1909,8 @@ def integration_diff(**kwargs):
     """
     Checks for differences between two versions of an integration, and verified that the new version covered the old version.
     """
-
+    from demisto_sdk.commands.integration_diff.integration_diff_detector import \
+        IntegrationDiffDetector
     integration_diff_detector = IntegrationDiffDetector(
         new=kwargs.get('new', ''),
         old=kwargs.get('old', ''),
@@ -1891,6 +1942,7 @@ def convert(config, **kwargs):
     Convert the content of the pack/directory in the given input to be compatible with the version given by
     version command.
     """
+    from demisto_sdk.commands.convert.convert_manager import ConvertManager
     check_configuration_file('convert', kwargs)
     sys.path.append(config.configuration.env_dir)
 
@@ -1918,6 +1970,8 @@ def convert(config, **kwargs):
 )
 @pass_config
 def error_code(config, **kwargs):
+    from demisto_sdk.commands.error_code_info.error_code_info import \
+        generate_error_code_information
     check_configuration_file('error-code-info', kwargs)
     sys.path.append(config.configuration.env_dir)