+++ demisto_sdk/commands/common/hook_validations/incident_field.py
@@ -226,26 +226,10 @@ class IncidentFieldValidator(ContentEntityValidator):
             "tickets",
             "issues",
         }
-        # TODO remove when demisto/etc#24232 is resolved
-        whitelisted_field_names = {
-            "XDR Alert Count",
-            "XDR High Severity Alert Count",
-            "XDR Medium Severity Alert Count",
-            "XDR Low Severity Alert Count",
-            "XDR Incident ID",
-            "Detection Ticketed",
-            "Claroty Alert Resolved",  # Needed for incidentfield-Claroty_Alert_Resolved.json
-            "Claroty Alert Type",  # Needed for incidentfield-Claroty_Alert_Type.json
-            "Code42 Alert Type",  # Needed for incidentfield-Code42_Alert_Type.json
-            "Code42 File Events",  # Needed for incidentfield-Code42_File_Events.json
-            "XDR Alerts",  # Needed for XDR_Alerts.json
-            "Indeni Issue ID",  # Needed for incidentfield-Indeni_Device_ID.json
-        }
         found_words = []
-        if name not in whitelisted_field_names:
-            for word in name.split():
-                if word.lower() in bad_words:
-                    found_words.append(word)
+        for word in name.split():
+            if word.lower() in bad_words:
+                found_words.append(word)
 
         if found_words:
             error_message, error_code = Errors.invalid_incident_field_name(found_words)