+++ demisto_sdk/commands/common/tests/update_id_set_test.py
@@ -8,6 +8,7 @@ from pathlib import Path
 
 import pytest
 
+import demisto_sdk.commands.common.update_id_set as uis
 from demisto_sdk.commands.common.constants import (DEFAULT_JOB_FROM_VERSION,
                                                    JOBS_DIR, FileType)
 from demisto_sdk.commands.common.legacy_git_tools import git_path
@@ -23,7 +24,7 @@ from demisto_sdk.commands.common.update_id_set import (
     get_report_data, get_script_data, get_values_for_keys_recursively,
     get_widget_data, has_duplicate, merge_id_sets, process_general_items,
     process_incident_fields, process_integration, process_jobs, process_script,
-    re_create_id_set)
+    re_create_id_set, should_skip_item_by_mp)
 from TestSuite.utils import IsEqualFunctions
 
 TESTS_DIR = f'{git_path()}/demisto_sdk/tests'
@@ -37,7 +38,8 @@ class TestPacksMetadata:
         'author': 'Cortex XSOAR',
         'tags': ['Alerts'],
         'useCases': ['Case Management'],
-        'categories': ['Endpoint']
+        'categories': ['Endpoint'],
+        'marketplaces': ['xsoar', 'marketplacev2'],
     }
 
     METADATA_WITH_PARTNER_SUPPORT = {
@@ -47,7 +49,8 @@ class TestPacksMetadata:
         'author': 'Some Partner',
         'tags': ['Alerts'],
         'useCases': ['Case Management'],
-        'categories': ['Endpoint']
+        'categories': ['Endpoint'],
+        'marketplaces': ['xsoar', 'marketplacev2'],
     }
 
     METADATA_WITH_COMMUNITY_SUPPORT = {
@@ -57,7 +60,8 @@ class TestPacksMetadata:
         'author': 'Someone',
         'tags': ['Alerts'],
         'useCases': ['Case Management'],
-        'categories': ['Endpoint']
+        'categories': ['Endpoint'],
+        'marketplaces': ['xsoar', 'marketplacev2'],
     }
 
     TEST_PACK = [
@@ -77,8 +81,8 @@ class TestPacksMetadata:
         Then
             - parsing all the data from file successfully
         """
-        import demisto_sdk.commands.common.update_id_set as uis
         mocker.patch.object(uis, 'get_pack_name', return_value='Pack1')
+        mocker.patch.object(uis, 'get_mp_types_from_metadata_by_item', return_value=['xsoar', 'marketplacev2'])
 
         pack = repo.create_pack("Pack1")
         pack.pack_metadata.write_json(metadata_file_content)
@@ -94,6 +98,23 @@ class TestPacksMetadata:
         assert result.get('tags') == ['Alerts']
         assert result.get('use_cases') == ['Case Management']
         assert result.get('categories') == ['Endpoint']
+        assert result.get('marketplaces') == ['xsoar', 'marketplacev2']
+
+    @staticmethod
+    def test_process_metadata__marketplace_mismatch(mocker):
+        """
+        Given
+            - Pack metadata
+
+        When
+            - parsing pack files when there is a mismatch between the item's marketplaces and the current run marketplace.
+
+        Then
+            - return empty dict
+        """
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=True)
+        res = get_pack_metadata_data('', print_logs=False)
+        assert res == {}
 
     @staticmethod
     @pytest.mark.parametrize('print_logs', [True, False])
@@ -107,8 +128,8 @@ class TestPacksMetadata:
         Then
             - Verify output to logs.
         """
-        import demisto_sdk.commands.common.update_id_set as uis
         mocker.patch.object(uis, 'get_pack_name', return_value='Pack1')
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
 
         pack = repo.create_pack("Pack1")
         pack.pack_metadata.write_json({
@@ -118,7 +139,8 @@ class TestPacksMetadata:
             'support': 'xsoar',
             'tags': ['Alerts'],
             'useCases': ['Case Management'],
-            'categories': ['Endpoint']
+            'categories': ['Endpoint'],
+            'marketplaces': ['xsoar', 'marketplacev2']
         })
         pack_metadata_path = pack.pack_metadata.path
         res = get_pack_metadata_data(pack_metadata_path, print_logs)
@@ -132,11 +154,12 @@ class TestPacksMetadata:
         assert res['Pack1']['use_cases'] == ['Case Management']
         assert res['Pack1']['categories'] == ['Endpoint']
         assert res['Pack1']['certification'] == 'certified'
+        assert res['Pack1']['marketplaces'] == ['xsoar', 'marketplacev2']
 
         assert (f'adding {pack_metadata_path} to id_set' in captured.out) == print_logs
 
     @staticmethod
-    def test_process_packs_exception_thrown(capsys):
+    def test_process_packs_exception_thrown(capsys, mocker):
         """
         Given
             - A pack metadata file path.
@@ -145,7 +168,7 @@ class TestPacksMetadata:
         Then
             - Handle the exceptions gracefully.
         """
-
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
         with pytest.raises(FileNotFoundError):
             get_pack_metadata_data('Pack_Path', True)
         captured = capsys.readouterr()
@@ -381,7 +404,7 @@ class TestIntegrations:
         }
     }
 
-    def test_process_integration__sanity(self):
+    def test_process_integration__sanity(self, mocker):
         """
         Given
             - A valid script package folder located at Packs/DummyPack/Scripts/DummyScript.
@@ -394,7 +417,7 @@ class TestIntegrations:
         """
         non_unified_file_path = os.path.join(TESTS_DIR, 'test_files',
                                              'DummyPack', 'Integrations', 'DummyIntegration')
-
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
         res = process_integration(non_unified_file_path, True)
         assert len(res) == 1
         non_unified_integration_data = res[0]
@@ -420,7 +443,7 @@ class TestIntegrations:
             assert IsEqualFunctions.is_dicts_equal(returned_data, const_data)
 
     @staticmethod
-    def test_process_integration__exception():
+    def test_process_integration__exception(mocker):
         """
         Given
             - An invalid "integration" file located at invalid_file_structures where commonfields object is not a dict.
@@ -431,9 +454,27 @@ class TestIntegrations:
         Then
             - an exception will be raised
         """
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
         test_file_path = os.path.join(TESTS_DIR, 'test_files', 'invalid_file_structures', 'integration.yml')
         with pytest.raises(Exception):
             process_integration(test_file_path, True)
+
+    @staticmethod
+    def test_process_integration__marketplace_mismatch(mocker):
+        """
+        Given
+            - An integration file
+
+        When
+            - parsing item files when there is a mismatch between the item's marketplaces and the current run marketplace.
+
+        Then
+            - return empty list
+        """
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=True)
+        test_file_path = os.path.join(TESTS_DIR, 'test_files', 'invalid_file_structures', 'integration.yml')
+        res = process_integration(test_file_path, print_logs=False)
+        assert res == []
 
 
 class TestScripts:
@@ -476,7 +517,7 @@ class TestScripts:
         assert IsEqualFunctions.is_dicts_equal(returned_data, const_data)
 
     @staticmethod
-    def test_process_script__sanity_package():
+    def test_process_script__sanity_package(mocker):
         """
         Given
             - An invalid "script" file located at invalid_file_structures where commonfields object is not a dict.
@@ -489,6 +530,7 @@ class TestScripts:
         """
         test_file_path = os.path.join(TESTS_DIR, 'test_files',
                                       'Packs', 'DummyPack', 'Scripts', 'DummyScript')
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
         res = process_script(test_file_path, True)
         assert len(res) == 1
         data = res[0]
@@ -501,7 +543,25 @@ class TestScripts:
         assert IsEqualFunctions.is_dicts_equal(returned_data, const_data)
 
     @staticmethod
-    def test_process_script__exception():
+    def test_process_script__marketplace_mismatch(mocker):
+        """
+        Given
+            - A script file
+
+        When
+            - parsing script files when there is a mismatch between the item's marketplaces and the current run marketplace.
+
+        Then
+            - return empty list
+        """
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=True)
+        test_file_path = os.path.join(TESTS_DIR, 'test_files', 'invalid_file_structures', 'integration.yml')
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
+        res = process_script(test_file_path, print_logs=False)
+        assert res == []
+
+    @staticmethod
+    def test_process_script__exception(mocker):
         """
         Given
             - An invalid "script" file located at invalid_file_structures where commonfields object is not a dict.
@@ -513,6 +573,8 @@ class TestScripts:
             - an exception will be raised
         """
         test_file_path = os.path.join(TESTS_DIR, 'test_files', 'invalid_file_structures', 'script.yml')
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
+
         with pytest.raises(Exception):
             process_script(test_file_path, True)
 
@@ -580,7 +642,7 @@ class TestPlaybooks:
         assert 'indicator_fields' in result.keys()
 
     @staticmethod
-    def test_get_playbook_data_no_fields():
+    def test_get_playbook_data_no_fields(mocker):
         """
         Given
             - A playbook file called playbook-no-incident-fields.yml without any
@@ -593,6 +655,8 @@ class TestPlaybooks:
             - parsing all the data from file successfully
         """
         test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/playbook-no-incident-fields.yml'
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
+
         result = get_playbook_data(test_dir)
         result = result.get('Arcsight - Get events related to the Case')
         assert 'name' in result.keys()
@@ -604,7 +668,7 @@ class TestPlaybooks:
         assert 'indicator_fields' not in result.keys()
 
     @staticmethod
-    def test_process_playbook__exception():
+    def test_process_playbook__exception(mocker):
         """
         Given
             - An invalid "playbook" file located at invalid_file_structures where tasks object is not a dict.
@@ -615,6 +679,7 @@ class TestPlaybooks:
         Then
             - an exception will be raised
         """
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
         test_file_path = os.path.join(TESTS_DIR, 'test_files', 'invalid_file_structures', 'playbook.yml')
         with pytest.raises(Exception):
             process_general_items(test_file_path, True, (FileType.PLAYBOOK,), get_playbook_data)
@@ -757,7 +822,7 @@ class TestPlaybooks:
         assert filters == ['isEqualString']
 
     @staticmethod
-    def test_get_transformers_from_playbook_inputs():
+    def test_get_transformers_from_playbook_inputs(mocker):
         """
         Given
         - playbook with 2 inputs that using Length and toUpperCase transformers
@@ -773,6 +838,8 @@ class TestPlaybooks:
                                                                    ]}}},
                            {'value': {'complex': {'transformers': [{'operator': 'Length'}
                                                                    ]}}}]}
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
+
         transformers, _ = get_filters_and_transformers_from_playbook(data)
         assert len(transformers) == 2
         assert 'toUpperCase' in transformers
@@ -781,7 +848,7 @@ class TestPlaybooks:
 
 class TestLayouts:
     @staticmethod
-    def test_process_layouts__sanity():
+    def test_process_layouts__sanity(mocker):
         """
         Given
             - A layout file called layout-to-test.json
@@ -794,6 +861,8 @@ class TestLayouts:
         """
         test_file = os.path.join(git_path(), 'demisto_sdk', 'commands', 'create_id_set', 'tests',
                                  'test_data', 'layout-to-test.json')
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
+
         res = process_general_items(test_file, True, (FileType.LAYOUT,), get_layout_data)
         assert len(res) == 1
         result = res[0]
@@ -808,7 +877,7 @@ class TestLayouts:
         assert 'incident_and_indicator_fields' in result.keys()
 
     @staticmethod
-    def test_process_layouts__no_incident_types_and_fields():
+    def test_process_layouts__no_incident_types_and_fields(mocker):
         """
         Given
             - A layout file called layout-to-test.json that doesnt have related incident fields and indicator fields
@@ -821,6 +890,7 @@ class TestLayouts:
         """
         test_file = os.path.join(git_path(), 'demisto_sdk', 'commands', 'create_id_set', 'tests',
                                  'test_data', 'layout-to-test-no-types-fields.json')
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
 
         res = process_general_items(test_file, False, (FileType.LAYOUT,), get_layout_data)
         assert len(res) == 1
@@ -836,7 +906,7 @@ class TestLayouts:
         assert 'incident_and_indicator_fields' not in result.keys()
 
     @staticmethod
-    def test_process_layoutscontainer__sanity():
+    def test_process_layoutscontainer__sanity(mocker):
         """
         Given
             - A layoutscontainer file called layoutscontainer-to-test.json
@@ -847,6 +917,7 @@ class TestLayouts:
         """
         test_file = os.path.join(git_path(), 'demisto_sdk', 'commands', 'create_id_set', 'tests',
                                  'test_data', 'layoutscontainer-to-test.json')
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
 
         res = process_general_items(test_file, True, (FileType.LAYOUTS_CONTAINER,), get_layoutscontainer_data)
         assert len(res) == 1
@@ -864,7 +935,7 @@ class TestLayouts:
 
 class TestIncidentFields:
     @staticmethod
-    def test_process_incident_fields__sanity():
+    def test_process_incident_fields__sanity(mocker):
         """
         Given
             - An incident field file called incidentfield-to-test.json
@@ -877,6 +948,8 @@ class TestIncidentFields:
         """
         test_dir = os.path.join(git_path(), 'demisto_sdk', 'commands', 'create_id_set', 'tests',
                                 'test_data', 'incidentfield-to-test.json')
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
+
         res = process_incident_fields(test_dir, True, [])
         assert len(res) == 1
         result = res[0]
@@ -889,7 +962,7 @@ class TestIncidentFields:
         assert 'scripts' in result.keys()
 
     @staticmethod
-    def test_process_incident_fields__no_types_scripts():
+    def test_process_incident_fields__no_types_scripts(mocker):
         """
         Given
             - An incident field file called incidentfield-to-test-no-types_scripts.json with no script or incident type
@@ -903,6 +976,8 @@ class TestIncidentFields:
         """
         test_dir = os.path.join(git_path(), 'demisto_sdk', 'commands', 'create_id_set', 'tests',
                                 'test_data', 'incidentfield-to-test-no-types_scripts.json')
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
+
         res = process_incident_fields(test_dir, True, [])
         assert len(res) == 1
         result = res[0]
@@ -914,10 +989,26 @@ class TestIncidentFields:
         assert 'incident_types' not in result.keys()
         assert 'scripts' not in result.keys()
 
+    @staticmethod
+    def test_process_incident_fields__marketplace_mismatch(mocker):
+        """
+        Given
+            - An incident field file
+        When
+            - parsing An incident field files when there is a mismatch between the item's marketplaces and the current run marketplace.
+        Then
+            - return empty list
+        """
+        test_dir = os.path.join(git_path(), 'demisto_sdk', 'commands', 'create_id_set', 'tests',
+                                'test_data', 'incidentfield-to-test-no-types_scripts.json')
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=True)
+        res = process_incident_fields(test_dir, True, [])
+        assert res == []
+
 
 class TestIndicatorType:
     @staticmethod
-    def test_process_indicator_type__sanity():
+    def test_process_indicator_type__sanity(mocker):
         """
         Given
             - An indicator type file called reputation-indicatortype.json.
@@ -929,6 +1020,7 @@ class TestIndicatorType:
             - parsing all the data from file successfully
         """
         test_dir = f'{git_path()}/demisto_sdk/commands/create_id_set/tests/test_data/reputation-indicatortype.json'
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
 
         result = get_indicator_type_data(test_dir, [{'integration': {'commands': ['ip']}}])
         result = result.get('indicator-type-dummy')
@@ -1016,7 +1108,7 @@ class TestIncidentTypes:
 
 class TestClassifiers:
     @staticmethod
-    def test_process_classifiers__no_types_scripts():
+    def test_process_classifiers__no_types_scripts(mocker):
         """
         Given
             - An classifier file called classifier-to-test-no-incidenttypes.json with incident type
@@ -1030,6 +1122,7 @@ class TestClassifiers:
         """
         test_file = os.path.join(git_path(), 'demisto_sdk', 'commands', 'create_id_set', 'tests',
                                  'test_data', 'classifier-to-test-no-incidenttypes.json')
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
 
         res = get_classifier_data(test_file)
         result = res.get('dummy classifier')
@@ -1661,7 +1754,7 @@ class TestReport:
 
 class TestGenericFunctions:
     @staticmethod
-    def test_process_general_items__sanity():
+    def test_process_general_items__sanity(mocker):
         """
         Given
             - A classifier file called classifier-to-test.json
@@ -1674,6 +1767,7 @@ class TestGenericFunctions:
         """
         test_file = os.path.join(git_path(), 'demisto_sdk', 'commands', 'create_id_set', 'tests',
                                  'test_data', 'classifier-to-test.json')
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
 
         res = process_general_items(test_file, True, (FileType.CLASSIFIER,), get_classifier_data)
         assert len(res) == 1
@@ -1688,7 +1782,25 @@ class TestGenericFunctions:
         assert 'dummy incident type 3' in result['incident_types']
 
     @staticmethod
-    def test_process_general_items__exception():
+    def test_process_general_items__marketplace_mismatch(mocker):
+        """
+        Given
+            - An item file
+
+        When
+            - parsing item files when there is a mismatch between the item's marketplaces and the current run marketplace.
+
+        Then
+            - return empty list
+        """
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=True)
+        test_file = os.path.join(git_path(), 'demisto_sdk', 'commands', 'create_id_set', 'tests',
+                                 'test_data', 'classifier-to-test.json')
+        res = process_general_items(test_file, True, (FileType.CLASSIFIER,), get_classifier_data)
+        assert res == []
+
+    @staticmethod
+    def test_process_general_items__exception(mocker):
         """
         Given
             - An invalid "dashboard" file located at invalid_file_structures where id is a list so it can't be used
@@ -1701,6 +1813,8 @@ class TestGenericFunctions:
             - an exception will be raised
         """
         test_file_path = os.path.join(TESTS_DIR, 'test_files', 'invalid_file_structures', 'dashboard.json')
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
+
         with pytest.raises(Exception):
             process_general_items(test_file_path, True, (FileType.DASHBOARD,), get_general_data)
 
@@ -2190,7 +2304,7 @@ class TestJob:
     @staticmethod
     @pytest.mark.parametrize('print_logs', (True, False))
     @pytest.mark.parametrize('is_feed', (True, False))
-    def test_process_jobs(capsys, repo, is_feed: bool, print_logs: bool):
+    def test_process_jobs(capsys, repo, is_feed: bool, print_logs: bool, mocker):
         """
         Given
             - A repo with a job object.
@@ -2202,6 +2316,7 @@ class TestJob:
         """
         pack = repo.create_pack()
         job_details = 'job details'
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
         job = pack.create_job(is_feed, details=job_details)
         res = process_jobs(job.path, print_logs)
 
@@ -2226,7 +2341,7 @@ class TestJob:
 
     @staticmethod
     @pytest.mark.parametrize('is_feed', (True, False))
-    def test_process_jobs_non_job_extension(capsys, repo, is_feed: bool):
+    def test_process_jobs_non_job_extension(capsys, repo, is_feed: bool, mocker):
         """
         Given
             - A file that isn't a valid Job (wrong filetype)
@@ -2236,6 +2351,8 @@ class TestJob:
         Then
             - Verify output to logs.
         """
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
+
         pack = repo.create_pack()
         job = pack.create_job(is_feed)
         job_path = Path(job.path)
@@ -2246,7 +2363,7 @@ class TestJob:
     @staticmethod
     @pytest.mark.parametrize('print_logs', (True, False))
     @pytest.mark.parametrize('is_feed', (True, False))
-    def test_process_jobs_file_nonexistent(capsys, repo, is_feed: bool, print_logs: bool):
+    def test_process_jobs_file_nonexistent(capsys, repo, is_feed: bool, print_logs: bool, mocker):
         """
         Given
             - A file that isn't a valid Job (missing file)
@@ -2256,6 +2373,8 @@ class TestJob:
         Then
             - Verify output to logs.
         """
+        mocker.patch.object(uis, 'should_skip_item_by_mp', return_value=False)
+
         pack = repo.create_pack()
         job = pack.create_job(is_feed)
         job_json_path = Path(job.path)
@@ -2514,3 +2633,24 @@ def test_get_filters_and_transformers_from_complex_value():
     assert 'Length' in transformers
     assert 'isEqualString' in filters
     assert 'StringContainsArray' in filters
+
+
+def test_should_skip_item_by_mp(mocker):
+    """
+    Given
+    - path of content item, the current marketplace this id set is generated for.
+    When
+    - when creating the id set, checking a content item that is 'xsoar only' if it should be part of the mpV2
+     id set.
+    Then
+    - return True since this item should be skipped.
+
+    """
+    import demisto_sdk.commands.common.update_id_set as uis
+    mocker.patch.object(uis, 'get_mp_types_from_metadata_by_item', return_value=['xsoar'])
+    pack_path = os.path.join(TESTS_DIR, 'test_files', 'DummyPackXsoarMPOnly')
+    script_path = os.path.join(TESTS_DIR, 'test_files', 'DummyPackScriptIsXsoarOnly', 'Scripts', 'DummyScript')
+    res1 = should_skip_item_by_mp(pack_path, 'mpv2')
+    res2 = should_skip_item_by_mp(script_path, 'mpv2')
+    assert res1
+    assert res2